requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
flask
flask-cors
python-dotenv
Pillow
flask-swagger-ui
fastapi
uvicorn